Created and deleted several components...
Deleted dummy DefaultCity.js and included this code as props in Weather.js (now SearchEngine.js) from which the whole weather data portion was cut out form the new component WeatherData.js. Also, CurrentDate.js was created to handle the date code and WeatherIcon to handle the icon code. A TemperatureCelsius.js component was created as yet just a dummy to handle the temperature calculation to Fahrenheit later.